+++ keja-backend/graphql/resolvers/bookingResolver.js
@@ -1,0 +1,54 @@
+const Home = require('../../models/home');
+const Booking = require('../../models/bookings');
+const { singleHome, user, transformHome} = require('./merge');
+
+const bookingResolver = {
+    bookings: async () => {
+        try{
+            const bookings = await Booking.find();
+            return bookings.map(booking => {
+                return {
+                    ...booking._doc, 
+                    _id: booking.id,
+                    user: user.bind(this, booking._doc.user),
+                    home: singleHome.bind(this, booking._doc.home)
+                    // createdAt: new Date(booking._doc.createdAt).toISOString, 
+                    // updatedAt: new Date(booking._doc.updatedAt).toISOString
+                }
+            });
+        } catch(err){
+            throw err;
+        }
+    },
+
+    bookHome: async args => {
+        const fetchedHome = await Home.findOne({_id: args.homeId});
+       const booking = new Booking({
+           user:'5d1475e80b6be54627b116a3',
+           home: fetchedHome
+       }); 
+       const result = await booking.save();
+       return {
+           ...result._doc, 
+           _id: result.id,
+           user: user.bind(this, booking._doc.user),
+           home: singleHome.bind(this, booking._doc.home)
+        //    createdAt: new Date(result._doc.createdAt).toISOString, 
+        //    updatedAt: new Date(result._doc.updatedAt).toISOString
+        }
+    },
+
+    cancelBooking: async args => {
+        try{
+            const abooking = await Booking.findById(args.bookingId).populate('home');
+            const home = transformHome(abooking.home)
+            await Booking.deleteOne({_id: args.bookingId});
+            return home;
+        }catch (err){
+            throw err;
+        }
+    }
+};
+
+
+module.exports = bookingResolver;
